pyproject.toml
[tool.poetry]
name = "nonebot_plugin_session"
version = "0.0.1"
description = "Nonebot2 会话目标提取与会话id定义"
authors = ["meetwq <meetwq@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/noneplugin/nonebot-plugin-session"
repository = "https://github.com/noneplugin/nonebot-plugin-session"

[tool.poetry.dependencies]
python = "^3.8"
nonebot2 = { version = "^2.0.0-rc.1", extras = ["fastapi"] }

[tool.poetry.group.dev.dependencies]
black = "^22.8.0"

[tool.poetry.group.test.dependencies]
pytest-cov = "^4.0.0"
pytest-asyncio = "^0.20.0"
nonebug = "^0.3.0"
nonebot2 = { version = "^2.0.0-rc.1", extras = ["fastapi"] }

[tool.poetry.group.adapters]
optional = true

[tool.poetry.group.adapters.dependencies]
nonebot-adapter-onebot = "^2.2.2"
nonebot-adapter-console = "^0.3.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
